Add files via upload
This project centers on developing a system that facilitates different types of network attacks for cybersecurity testing purposes. The system can generate four distinct types of attacks, each of which has a detailed description for user comprehension. It scans and displays network IP addresses and a list of possible attack methods. Users can select a specific or random attack and choose a target from the discovered IPs. The system logs every attack choice, including the type of attack, execution time, and target IP, stored in a log file