Add files via upload
This project involved the creation of a comprehensive program for digital forensic analysis. Focused on automated hard disk drive (HDD) and memory investigations, it was designed to identify, extract, and display crucial data elements like network traffic or human-readable information. Additionally, the program integrates with the Volatility software for in-depth memory analysis and outputs detailed reports. It also features a dedicated function to install necessary forensic tools, ensuring smooth operation.