Add files via upload
This project involves creating a script for comprehensive network device mapping, identifying ports, services, and vulnerabilities. The user defines the network range, after which the program deploys tools like nmap and masscan for scanning and mapping purposes, storing the data in a newly created directory. The script also probes for network vulnerabilities, employing nmap, searchsploit, hydra, and medusa to identify security gaps, such as weak passwords. Finally, the scan summary and findings are presented to the user.